services/report_generator/pyproject.toml
[project]
name = "report-generator"
version = "1.0.0"
description = "Report Generator Service - Excel and PDF Output"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "openpyxl>=3.1.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.7.0",
    "reportlab>=4.0.0",
    "PyYAML>=6.0",
    "python-multipart>=0.0.12",
    "dhara_shared",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["routers*", "services*", "core*", "schemas*", "feasibility*"]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
addopts = "--import-mode=importlib"
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]

[tool.uv.sources]
dhara_shared = { workspace = true }

